+++ src/app/shared/_components/funnel/components/step/components/step.tsx
@@ -27,7 +27,7 @@ const Step: React.FC<IStep> = ({ children, name }) => {
   //!SECTION 메서드 영역
 
   if (step === name) {
-    return <>{children}</>;
+    return <div className="h-full w-full">{children}</div>;
   }
 
   return null;